requirements.txt
numpy==1.19.2
Flask==1.1.2
opencv_python==4.4.0.44
Pillow==8.1.1
